OBLS-421 create dashboard card component
