+++ src/screens/Dashboard/index.tsx
@@ -1,6 +1,5 @@
 import React, { useCallback } from 'react';
 import { FlatList, ListRenderItemInfo, Text, View } from 'react-native';
-import { Card } from 'react-native-paper';
 import { useSelector } from 'react-redux';
 
 import EmptyIcon from '../../assets/images/icon_empty.svg';
@@ -8,8 +7,8 @@ import Button from '../../components/Button';
 import { useFilteredDashboardEntries } from '../../hooks/useFilteredDashboardEntries';
 import { useOrderedDashboardEntries } from '../../hooks/useOrderedDashboardEntries';
 import { useResponsiveColumns } from '../../hooks/useResponsiveColumns';
-import { navigateToDashboardEntry } from '../../NavigationService';
 import { RootState } from '../../redux/reducers';
+import { DashboardCard } from './DashboardCard';
 import { DashboardEntry } from './dashboardData';
 import styles from './styles';
 import type { Props } from './Types';
@@ -23,21 +22,7 @@ export default function Dashboard({ navigation }: Props) {
   const visibleEntries = useFilteredDashboardEntries(orderedEntries, dashboardEntriesVisibility);
   const { columns } = useResponsiveColumns();
 
-  const renderItem = useCallback(({ item }: ListRenderItemInfo<DashboardEntry>) => {
-    const IconComponent = item.icon;
-    const onPress = () => navigateToDashboardEntry(item);
-
-    return (
-      <Card style={styles.cardContainer} onPress={onPress}>
-        <Card.Content style={styles.cardContent}>
-          <View style={styles.iconWrapper}>
-            {IconComponent && <IconComponent width={styles.icon.width} height={styles.icon.height} />}
-          </View>
-          <Text style={styles.cardLabel}>{item.screenName}</Text>
-        </Card.Content>
-      </Card>
-    );
-  }, []);
+  const renderItem = useCallback(({ item }: ListRenderItemInfo<DashboardEntry>) => <DashboardCard item={item} />, []);
 
   return (
     <View style={styles.screenContainer}>
